move get_communicator_class under communicator
